Move sync examples to a separate package
